Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ int main() {
     float d;
 
     scanf("%d %d %d", &a, &b, &c);
-    d=a+b+c/3;
+    d=float((a+b+c)/3);
     printf("Average: %.2f",d);
     return 0;
 }
